Move page_regex util around
